1. Container
To center content in the body of the page we'll need to create a container that is positioned and styled in CSS and then this class will be applied to several HTML elements in our index.html.

To start, in the css/main.css file create a .container class that sets the top and bottom margins to 0 and the left and right to auto.

Set the max-width to 1170px.