10. Buttons
Each button will have a consistent look and feel. In main.css, add the class .btn with the following styles:

    add a cursor
    remove any borders
    padding should be 6px top and bottom and 12px left and right
    Round the corners by 3px
    change the line height to 1.4

Also, Create two different button styles btn-default and btn-primary. The color of the text for both should be white. The btn-default background color is #faa541. The btn-primary background color should be #364147.