12. Form Controls
We only have a few form controls, but let's make a custom class for them anyway. Add a new class called .form-control below the button classes in main.css.

This class should change the display to block, and add padding of 6px top/bottom and 12px left/right. Next, add a 1px solid border with a stroke color of #a0a0a0. The text in the control should have a color of #8598a2. Finally, add a line height of 1.4.