14. Typography
There are some base typographic styles we'll use throughout the page.

In the main stylesheet, below the .form-control class, add three classes: .text-light, .text-secondary, and .text-primary.

The light and secondary classes should have a font size of 30px. The light color is #a0a0a0 and the secondary color is #faa541.

The primary class should have a font size of 16px and a color of #364147.