11. Profile
To style the profile section of the header we need two different classes. Name them .profile and .account. The only declaration in the .profile class should set the left margin to auto. The .account class should adjust the alignment on the text to the right, have a full padding of 15px, and a font color and size of #364147 and 18px.

Finally, select all the links that are descendant of all elements with a class of .account. Overwrite the font-size, set it to 14px.