12. Profile Picture
We'll round the profile picture and size it down.

Create a class called .profile-pic in main.css.

Make the size of the profile picture container 60x60.

Then create a rule that selects the all descendant img elements of .profile-pic.

To round the image set the border-radius to 100px. Make sure the image fits in the container set max-width to 100%.