13. Profile Class Attribute
We have three classes to apply in index.html - .profile, .account, and .profile-pic.

In the <!-- PROFILE --> block assign the first <div> the class profile. Add the account class to the nested <div>. Finally, give the <div> that surrounds the image with an src of img/profile-pic.jpg the profile-pic class.